refactor(SDK): remove SDK.get in favor of direct import
