+++ plugins/tracking/__tests__/TrackingPlugin-test.js
@@ -9,11 +9,11 @@ jest.setMock("react-router", {
 });
 
 const PluginTestUtils = require("PluginTestUtils");
+const DOMUtils = require("#SRC/js/utils/DOMUtils");
 
 const SDK = PluginTestUtils.getSDK("tracking", { enabled: true });
 require("../SDK").setSDK(SDK);
 
-const DOMUtils = SDK.get("DOMUtils");
 const TrackingHooks = require("../hooks");
 
 describe("TrackingHooks", function() {
