--- src/js/plugin-bridge/PluginModules.js
@@ -1,117 +1,0 @@
-// Modules to paths
-module.exports = {
-  constants: {
-    EventTypes: "EventTypes",
-    HTTPStatusCodes: "HTTPStatusCodes",
-    TransactionTypes: "TransactionTypes",
-    UserActions: "UserActions"
-  },
-  events: {
-    SidebarActions: "SidebarActions"
-  },
-  stores: {
-    DCOSStore: "DCOSStore",
-    AuthStore: "AuthStore",
-    BaseStore: "BaseStore",
-    ConfigStore: "ConfigStore",
-    CosmosPackagesStore: "CosmosPackagesStore",
-    MesosStateStore: "MesosStateStore",
-    MetadataStore: "MetadataStore",
-    NotificationStore: "NotificationStore",
-    UsersStore: "UsersStore",
-    UserStore: "UserStore",
-    VisibilityStore: "VisibilityStore"
-  },
-  structs: {
-    Batch: "Batch",
-    CompositeState: "CompositeState",
-    Item: "Item",
-    List: "List",
-    Transaction: "Transaction",
-    Tree: "Tree"
-  },
-  utils: {
-    ApplicationUtil: "ApplicationUtil",
-    CookieUtils: "CookieUtils",
-    DateUtil: "DateUtil",
-    DOMUtils: "DOMUtils",
-    FormUtil: "FormUtil",
-    LocalStorageUtil: "LocalStorageUtil",
-    Maths: "Maths",
-    ParserUtil: "ParserUtil",
-    ReducerUtil: "ReducerUtil",
-    ResourceTableUtil: "ResourceTableUtil",
-    RouterUtil: "RouterUtil",
-    StringUtil: "StringUtil",
-    TableUtil: "TableUtil",
-    TabsUtil: "TabsUtil",
-    Units: "Units",
-    Util: "Util"
-  },
-  mixins: {
-    InternalStorageMixin: "InternalStorageMixin",
-    TabsMixin: "TabsMixin"
-  },
-  components: {
-    AccessDeniedPage: "AccessDeniedPage",
-    ActionsModal: "modals/ActionsModal",
-    AlertPanel: "AlertPanel",
-    AlertPanelHeader: "AlertPanelHeader",
-    Authenticated: "Authenticated",
-    Breadcrumbs: "Breadcrumbs",
-    Breadcrumb: "Breadcrumb",
-    BreadcrumbSupplementalContent: "BreadcrumbSupplementalContent",
-    BreadcrumbTextContent: "BreadcrumbTextContent",
-    Chart: "charts/Chart",
-    CheckboxTable: "CheckboxTable",
-    ClipboardTrigger: "ClipboardTrigger",
-    ClusterName: "ClusterName",
-    ConfigurationMap: "ConfigurationMap",
-    ConfigurationMapHeading: "ConfigurationMapHeading",
-    ConfigurationMapLabel: "ConfigurationMapLabel",
-    ConfigurationMapRow: "ConfigurationMapRow",
-    ConfigurationMapSection: "ConfigurationMapSection",
-    ConfigurationMapValue: "ConfigurationMapValue",
-    CollapsingString: "CollapsingString",
-    CreateServiceModalServicePickerOption: "CreateServiceModalServicePickerOption",
-    CreateServiceModalServicePickerOptionContent: "CreateServiceModalServicePickerOptionContent",
-    CreateServiceModalServicePickerOptionImage: "CreateServiceModalServicePickerOptionImage",
-    DetailViewHeader: "DetailViewHeader",
-    DetailViewSectionHeading: "DetailViewSectionHeading",
-    DCOSLogo: "DCOSLogo",
-    FieldAutofocus: "form/FieldAutofocus",
-    FilterBar: "FilterBar",
-    FilterButtons: "FilterButtons",
-    FilterHeadline: "FilterHeadline",
-    FilterInputText: "FilterInputText",
-    FluidGeminiScrollbar: "FluidGeminiScrollbar",
-    FormModal: "FormModal",
-    HashMapDisplay: "HashMapDisplay",
-    Icon: "Icon",
-    Image: "Image",
-    Loader: "Loader",
-    MesosphereLogo: "icons/MesosphereLogo",
-    ModalHeading: "modals/ModalHeading",
-    NestedServiceLinks: "NestedServiceLinks",
-    Page: "Page",
-    Panel: "Panel",
-    RequestErrorMsg: "RequestErrorMsg",
-    StatusBar: "StatusBar",
-    TabForm: "TabForm",
-    ToggleButton: "ToggleButton",
-    Typeahead: "Typeahead",
-    UserAccountDropdown: "UserAccountDropdown",
-    UserAccountDropdownTrigger: "UserAccountDropdownTrigger",
-    UserFormModal: "modals/UserFormModal",
-    "form-elements": "form/index"
-  },
-  config: {
-    Config: "Config"
-  },
-  systemPages: {
-    UsersPage: "UsersPage"
-  },
-  internalPlugin: {
-    services: "services/index"
-  }
-};
